feat : randomRecoList -> popRecoList 로 변경
